+++ src/components/About.astro
@@ -1,45 +1,9 @@
 ---
-import authorImg from "../assets/page/author.png";
-import GithubStats from "./GithubStats.astro";
-import siteConfig from '../content/site-config';
-
-const socialLinks = siteConfig.socialLinks;
+import AboutLayout from "../custom/AboutLayout.astro"
 ---
 
 <section class="mt-12 bg-gray-50 dark:bg-gray-800 rounded-xl p-6">
   <h2 class="text-2xl font-bold text-gray-900 dark:text-white mb-4">About Me</h2>
-  <div class="flex flex-col md:flex-row gap-6">
-    <div class="md:w-1/3 mb-8 flex justify-center">
-      <img 
-        src={authorImg.src} 
-        alt="Syam's profile picture" 
-        class="rounded-2xl w-48 h-48 object-cover border-4 border-white dark:border-gray-700 shadow-lg"
-      />
-    </div>
-    <div class="md:w-2/3 mb-8">
-      <p class="text-gray-700 dark:text-gray-300 mb-4">
-        {siteConfig.authorDescription}
-      </p>
-      {socialLinks.length > 0 && (
-        <div class="flex space-x-4">
-          {socialLinks.map((link) => (
-            <a 
-              href={link.url} 
-              target="_blank" 
-              class="text-blue-600 hover:text-blue-800 dark:text-blue-400 dark:hover:text-blue-300 font-medium"
-            >
-              {link.name}
-            </a>
-          ))}
-        </div>
-      )}
-    </div>
-  </div>
 
-  <hr>
-
-  <GithubStats />
-
-  <!-- You can add more content here -->
-
+  <AboutLayout />
 </section>
