+++ src/components/Footer.astro
@@ -1,5 +1,5 @@
 ---
-import siteConfig from '../content/site-config';
+import siteConfig from '../custom/site-config';
 ---
 
 <footer class="mt-12 py-8 text-center text-gray-600 dark:text-gray-300 text-sm border-t border-gray-200 dark:border-gray-700 backdrop-blur-sm bg-white/50 dark:bg-gray-900/50 relative">
