+++ src/components/Header.astro
@@ -1,6 +1,6 @@
 ---
 import "../styles/home.css";
-import siteConfig from '../content/site-config';
+import siteConfig from '../custom/site-config';
 ---
 
 <header class="relative overflow-hidden bg-gradient-to-r from-blue-500 to-purple-600 text-white">
