+++ src/components/Hero.astro
@@ -1,5 +1,5 @@
 ---
-import siteConfig from '../content/site-config';
+import siteConfig from '../custom/site-config';
 ---
 
 <section class="py-8 text-center border-b border-gray-200 dark:border-gray-700">
