+++ src/components/Navbar.astro
@@ -1,6 +1,6 @@
 ---
 import "../styles/home.css";
-import siteConfig from '../content/site-config';
+import siteConfig from '../custom/site-config';
 import blogImg from "../assets/page/blog.png";
 
 interface Props {
