+++ src/custom/AboutLayout.astro
@@ -1,0 +1,12 @@
+---
+import Introduce from "./Introduce.astro"
+import GithubStats from "../components/GithubStats.astro"
+---
+
+<Introduce />
+
+<hr>
+
+<GithubStats />
+
+<!-- You can add more content and arrange here -->
