+++ src/custom/Introduce.astro
@@ -1,0 +1,34 @@
+---
+import authorImg from "../assets/page/author.png";
+import siteConfig from "./site-config.ts";
+
+const socialLinks = siteConfig.socialLinks;
+---
+
+<div class="flex flex-col md:flex-row gap-6">
+  <div class="md:w-1/3 mb-8 flex justify-center">
+    <img 
+      src={authorImg.src} 
+      alt="Syam's profile picture" 
+      class="rounded-2xl w-48 h-48 object-cover border-4 border-white dark:border-gray-700 shadow-lg"
+    />
+  </div>
+  <div class="md:w-2/3 mb-8">
+    <p class="text-gray-700 dark:text-gray-300 mb-4">
+      {siteConfig.authorDescription}
+    </p>
+    {socialLinks.length > 0 && (
+      <div class="flex space-x-4">
+        {socialLinks.map((link) => (
+          <a 
+            href={link.url} 
+            target="_blank" 
+            class="text-blue-600 hover:text-blue-800 dark:text-blue-400 dark:hover:text-blue-300 font-medium"
+          >
+            {link.name}
+          </a>
+        ))}
+      </div>
+    )}
+  </div>
+</div>
