+++ src/layouts/Layout.astro
@@ -1,7 +1,7 @@
 ---
 const { title } = Astro.props;
 import "../styles/bubbles.css"
-import siteConfig from '../content/site-config';
+import siteConfig from '../custom/site-config';
 ---
 
 <!doctype html>
