+++ src/layouts/PostLayout.astro
@@ -3,7 +3,7 @@ import "../styles/bubbles.css";
 import "../styles/page.css";
 import MarkdownRenderer from '../components/MarkdownRenderer.astro';
 import Footer from '../components/Footer.astro';
-import siteConfig from '../content/site-config';
+import siteConfig from '../custom/site-config';
 const { title, description, pubDate, content } = Astro.props;
 ---
