+++ src/layouts/RepoLayout.astro
@@ -8,7 +8,7 @@ interface Props {
 import "../styles/bubbles.css";
 import "../styles/page.css";
 import Footer from '../components/Footer.astro';
-import siteConfig from '../content/site-config';
+import siteConfig from '../custom/site-config';
 const { title, repoUrl, languages } = Astro.props as Props;
 ---
