+++ src/pages/index.astro
@@ -7,7 +7,7 @@ import FeaturedPosts from "../components/FeaturedPosts.astro";
 import About from "../components/About.astro";
 import Navbar from "../components/Navbar.astro";
 import Footer from "../components/Footer.astro";
-import siteConfig from "../content/site-config";
+import siteConfig from "../custom/site-config";
 ---
 
 <html lang="en" class="scroll-smooth">
